Split out cluster and clusterstore from secretcontroller
